✔ typed userMovieList slice
